Make TRACEN-X more modular
